chore: separate consensus worker logic from header validation.
