+++ amaru/src/consensus/mod.rs
@@ -1,5 +1,6 @@
 pub mod header_validation;
 pub mod nonce;
+pub mod worker;
 
 pub type RawBlock = Vec<u8>;
 pub type Point = pallas_network::miniprotocols::Point;
